Handle connection reset on handshake
The SSL socket performs some operations on the underlying socket when accept()
is called, even when do_handshake_on_connect is False. Those operations can
apparently result in a ConnectionResetError being raised in some circumstances.

In the specific scenario encountered, SSLSocket called getpeername() which
raised a ENOTCONNECTED error, then it perform a recv(1) which caused a
ConnectionResetError.